LayoutTests/fast/canvas/webgl/array-override-set-expected.txt

Test overriding the set method of WebGL array

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Regression test for ????
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS array[0] is 0
PASS successfullyParsed is true

TEST COMPLETE
